+++ pulldown-cmark-mdcat/src/resources/file.rs
@@ -9,7 +9,9 @@
 use std::fs::File;
 use std::io::prelude::*;
 use std::io::{Error, ErrorKind, Result};
+use std::path::Path;
 
+use mime::Mime;
 use tracing::{event, instrument, Level};
 use url::Url;
 
@@ -28,6 +30,27 @@ impl FileResourceHandler {
     pub fn new(read_limit: u64) -> Self {
         Self { read_limit }
     }
+}
+
+/// Guess a mimetype, in so far as mdcat makes use of the mime type.
+///
+/// This function recognizes
+///
+/// - SVG images because mdcat needs to render SVG images explicitly, and
+/// - PNG images because kitty can pass through PNG images in some cases.
+///
+/// It checks mime types exclusively by looking at the lowercase extension.
+///
+/// It ignores all other extensions and mime types and returns `None` in these cases.
+fn guess_mimetype<P: AsRef<Path>>(path: P) -> Option<Mime> {
+    path.as_ref()
+        .extension()
+        .map(|s| s.to_ascii_lowercase())
+        .and_then(|s| match s.to_str() {
+            Some("png") => Some(mime::IMAGE_PNG),
+            Some("svg") => Some(mime::IMAGE_SVG),
+            _ => None,
+        })
 }
 
 impl ResourceUrlHandler for FileResourceHandler {
@@ -54,7 +77,7 @@ impl ResourceUrlHandler for FileResourceHandler {
                             format!("Contents of {url} exceeded {} bytes", self.read_limit),
                         ))
                     } else {
-                        let mime_type = mime_guess::from_path(&path).first();
+                        let mime_type = guess_mimetype(&path);
                         if mime_type.is_none() {
                             event!(
                                 Level::DEBUG,