+++ CHANGELOG.md
@@ -8,6 +8,13 @@ Use `cargo release` to create a new release.
 
 ## [Unreleased]
 
+### Removed
+- Remove a few dependencies: `mime_guess` and `system_proxy` (see [GH-297]).
+- mdcat no longer supports building against the `onig` regex library (see [GH-297]).
+- mdcat no longer supports a static build against musl (see [GH-297]).
+
+[GH-297]: https://github.com/swsnr/mdcat/pull/297
+
 ## [2.5.0] – 2024-09-26
 
 ### Changed