+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Use `cargo release` to create a new release.
 
 ## [Unreleased]
 
+## [2.6.0] – 2024-11-13
+
 ### Added
 - Add support for images in Ghostty terminal (see [GH-300]).
 
@@ -809,7 +811,8 @@ Use `cargo release` to create a new release.
 - Support ordered and unordered lists, with nest.
 - Show links, with references grouped by section.
 
-[Unreleased]: https://github.com/swsnr/mdcat/compare/mdcat-2.5.0...HEAD
+[Unreleased]: https://github.com/swsnr/mdcat/compare/mdcat-2.6.0...HEAD
+[2.6.0]: https://github.com/swsnr/mdcat/compare/mdcat-2.5.0...mdcat-2.6.0
 [2.5.0]: https://github.com/swsnr/mdcat/compare/mdcat-2.4.0...mdcat-2.5.0
 [2.4.0]: https://github.com/swsnr/mdcat/compare/mdcat-2.3.1...mdcat-2.4.0
 [2.3.1]: https://github.com/swsnr/mdcat/compare/mdcat-2.3.0...mdcat-2.3.1